+++ libft/libft.h
@@ -71,7 +71,7 @@ char				*ft_strstr(const char *str, const char *to_find);
 char				*ft_strnstr(const char *s, const char *find, size_t len);
 int					ft_strcmp(const char *s1, const char *s2);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
-int					ft_atoi(const char *str);
+long long int		ft_atoi(const char *str);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
